pyproject.toml
[tool.poetry]
name = "fastapi-mongodb"
version = "0.0.1"
description = ""
authors = ["Hylan Silva <hylansilva@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pymongo = "^4.7.0"
motor = "^3.4.0"
fastapi = "0.111.0"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
pydantic-settings = "^2.2.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.9"

[tool.poetry.group.dev.dependencies]
taskipy = "^1.12.2"
pytest = "^8.2.0"
pytest-cov = "^5.0.0"
httpx = "^0.27.0"

[tool.taskipy.tasks]
dev = "fastapi dev app/main.py"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
